Update monitor stdlib package
Move Monitor integration tests from inline snippets executed in individual
Rust test functions to a single Ruby source file that executes all of the
tests as a group an 'asserts' by raising.

This pattern is the same as is used for testing the `InlineBuffer` Array
implementation.

vendorize monitor sources.
